Remove possibility to change a model's doc type field name
To allow having polymorphic documents in a single-type index, we added a field to the base `Model` class that declares what type the document has. That field is named "doctype" by default, but the possibility was left to override the name. However, leaving that possibility:
-  Makes the code quite a bit less clear, since it involves dynamically setting an attribute on the metaclass (see https://github.com/globality-corp/microcosm-elasticsearch/blob/970aa4c1c24d6f84edda39742659e7badfac6245/microcosm_elasticsearch/models.py#L9)
- Does not play nice with inheritance: the logic in the metaclass runs once for `Model` and each of its subclasses, when we'd like to only set the field in the leaf child class.
- Brings little value.

For those reasons, we've decided to simplify the code and always call that doc type field "doctype", without the ability to change it.