Allow running createall on existing index
With index.create, we get an error if the index already exists. This forces us to delete the index to update its mapping after a model change. Instead we can use index.save, which creates the index if it doesn't exist, and updates the mapping with the latest model definition otherwise.