GLOB-22921: Use the SigV4 logic from botocore instead of third-party libraries (#31)
Why?

Currently we are using `requests-aws4auth-redux`, a fork of `requests-aws4auth`. This was done to fix a problem with credentials expiring and needing to be refreshed, since we encountered the problem ~2 years ago, and `requests-aws4auth` had its last release ~5 months before we had that problem. Maintaining a fork of an unmaintained library to save ~15 lines of code seems like a bad trade.